fix price range when change opening position mode
